refactor: delete useless files
