--- src/features/Workflow/Dashboard/index.ts
@@ -1,1 +1,0 @@
-export * from "./Dashboard";
