--- src/features/Workflow/Ticket/index.ts
@@ -1,1 +1,0 @@
-export * from "./Ticket";
