--- src/features/Workflow/fixtures.ts
@@ -1,98 +1,0 @@
-export const data = [
-  {
-    id: 0,
-    priority: 0,
-    workflow: 2,
-    title: "Task 1",
-    description: "Descrierea taskului 1",
-    notes: "Note pentru task 1",
-  },
-  {
-    id: 1,
-    priority: 1,
-    workflow: 1,
-    title: "Task 2",
-    description: "Descrierea taskului 2",
-    notes: "Note pentru task 2",
-  },
-  {
-    id: 2,
-    priority: 2,
-    workflow: 0,
-    title: "Task 3",
-    description: "Descrierea taskului 3",
-    notes: "Note pentru task 3",
-  },
-  {
-    id: 3,
-    priority: 0,
-    workflow: 2,
-    title: "Task 4",
-    description: "Descrierea taskului 4",
-    notes: "Note pentru task 4",
-  },
-  {
-    id: 4,
-    priority: 3,
-    workflow: 1,
-    title: "Task 5",
-    description: "Descrierea taskului 5",
-    notes: "Note pentru task 5",
-  },
-  {
-    id: 5,
-    priority: 2,
-    workflow: 2,
-    title: "Task 6",
-    description: "Descrierea taskului 6",
-    notes: "Note pentru task 6",
-  },
-  {
-    id: 6,
-    priority: 1,
-    workflow: 0,
-    title: "Task 7",
-    description: "Descrierea taskului 7",
-    notes: "Note pentru task 7",
-  },
-  {
-    id: 7,
-    priority: 0,
-    workflow: 1,
-    title: "Task 8",
-    description: "Descrierea taskului 8",
-    notes: "Note pentru task 8",
-  },
-  {
-    id: 8,
-    priority: 2,
-    workflow: 2,
-    title: "Task 9",
-    description: "Descrierea taskului 9",
-    notes: "Note pentru task 9",
-  },
-  {
-    id: 9,
-    priority: 2,
-    workflow: 1,
-    title: "Task 10",
-    description: "Descrierea taskului 10",
-    notes: "Note pentru task 10",
-  },
-  {
-    id: 10,
-    priority: 1,
-    workflow: 0,
-    title: "Task 11",
-    description: "Descrierea taskului 11",
-    notes: "Note pentru task 11",
-  },
-  {
-    id: 11,
-    priority: 0,
-    workflow: 2,
-    title: "Task 12",
-    description: "Descrierea taskului 12",
-    notes: "Note pentru task 12",
-  },
-];
